+++ docs/changelog.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 Each revision is versioned by the date of the revision.
 
+## 2026-03-11
+
+- Build charms using the charmcraft wrapper.
+
 ## 2026-03-10
 
+- Drop support for charmcraftcache
 - Drop leftover registry input.
 
 ## 2026-03-09